+++ obtain_snp_dbs.sh
@@ -15,15 +15,18 @@ STANDARD_INDELS_LINK=https://storage.googleapis.com/gcp-public-data--broad-refer
 
 # The following for the human genome will take ~3 min
 # Obtain dbSNP v138 (1.45G)
-echo "Downloading dbSNP v138 and its index file (1.45G + 2.21M) ..."
-wget -O ${SNP_DB_DIR}/${DB_SNP_NAME} ${DB_SNP_LINK}
-# Obtain dbSNP v138 index file (2.21M)
-wget -O ${SNP_DB_DIR}/${DB_SNP_NAME}.tbi ${DB_SNP_LINK}.tbi
+if [ ! -f ${SNP_DB_DIR}/${DB_SNP_NAME} ]; then
+    echo "Downloading dbSNP v138 and its index file (1.45G + 2.21M) ..."
+    wget -O ${SNP_DB_DIR}/${DB_SNP_NAME} ${DB_SNP_LINK}
+    # Obtain dbSNP v138 index file (2.21M)
+    wget -O ${SNP_DB_DIR}/${DB_SNP_NAME}.tbi ${DB_SNP_LINK}.tbi
+fi
 
-# Obtain Mills and 1000G gold standard indels (19.73M)
-echo "Downloading Mills and 1000G gold standard indels and its index file (9.73M + 1.43M) ..."
-wget -O ${SNP_DB_DIR}/${INDELS_NAMES} ${STANDARD_INDELS_LINK}
-# Mills and 1000G gold standard indels index file (1.43M)
-wget -O ${SNP_DB_DIR}/${INDELS_NAMES}.tbi ${STANDARD_INDELS_LINK}.tbi
-
+if [ ! -f ${SNP_DB_DIR}/${INDELS_NAMES} ]; then
+    # Obtain Mills and 1000G gold standard indels (19.73M)
+    echo "Downloading Mills and 1000G gold standard indels and its index file (9.73M + 1.43M) ..."
+    wget -O ${SNP_DB_DIR}/${INDELS_NAMES} ${STANDARD_INDELS_LINK}
+    # Mills and 1000G gold standard indels index file (1.43M)
+    wget -O ${SNP_DB_DIR}/${INDELS_NAMES}.tbi ${STANDARD_INDELS_LINK}.tbi
+fi
